+++ src/ggcov-html.c
@@ -178,8 +178,8 @@ generate_stats(yaml_generator_t &yaml, const cov_stats_t &stats)
 
     yaml.begin_mapping();
 
-    yaml.key("blocks_executed").value(stats.blocks_executed());
-    yaml.key("blocks_total").value(stats.blocks_total());
+    yaml.key("blocks_executed").value((uint64_t)stats.blocks_executed());
+    yaml.key("blocks_total").value((uint64_t)stats.blocks_total());
     yaml.key("blocks_fraction").value(stats.blocks_fraction());
     yaml.key("blocks_percent").value((unsigned)(100.0*stats.blocks_fraction()+0.5));
     yaml.key("blocks_sort_fraction").value(stats.blocks_sort_fraction());
@@ -188,37 +188,37 @@ generate_stats(yaml_generator_t &yaml, const cov_stats_t &stats)
     yaml.key("blocks_abslabel").value(absbuf);
     yaml.key("blocks_pclabel").value(pcbuf);
 
-    yaml.key("lines_executed").value(stats.lines_executed());
-    yaml.key("lines_full").value(stats.lines_full());
-    yaml.key("lines_partial").value(stats.lines_partial());
-    yaml.key("lines_total").value(stats.lines_total());
+    yaml.key("lines_executed").value((uint64_t)stats.lines_executed());
+    yaml.key("lines_full").value((uint64_t)stats.lines_full());
+    yaml.key("lines_partial").value((uint64_t)stats.lines_partial());
+    yaml.key("lines_total").value((uint64_t)stats.lines_total());
     yaml.key("lines_fraction").value(stats.lines_fraction());
     yaml.key("lines_sort_fraction").value(stats.lines_sort_fraction());
     cov_stats_t::format_row_labels(stats.lines_by_status(), absbuf, pcbuf);
     yaml.key("lines_abslabel").value(absbuf);
     yaml.key("lines_pclabel").value(pcbuf);
 
-    yaml.key("functions_executed").value(stats.functions_executed());
-    yaml.key("functions_full").value(stats.functions_full());
-    yaml.key("functions_partial").value(stats.functions_partial());
-    yaml.key("functions_total").value(stats.functions_total());
+    yaml.key("functions_executed").value((uint64_t)stats.functions_executed());
+    yaml.key("functions_full").value((uint64_t)stats.functions_full());
+    yaml.key("functions_partial").value((uint64_t)stats.functions_partial());
+    yaml.key("functions_total").value((uint64_t)stats.functions_total());
     yaml.key("functions_fraction").value(stats.functions_fraction());
     yaml.key("functions_sort_fraction").value(stats.functions_sort_fraction());
     cov_stats_t::format_row_labels(stats.functions_by_status(), absbuf, pcbuf);
     yaml.key("functions_abslabel").value(absbuf);
     yaml.key("functions_pclabel").value(pcbuf);
 
-    yaml.key("calls_executed").value(stats.calls_executed());
-    yaml.key("calls_total").value(stats.calls_total());
+    yaml.key("calls_executed").value((uint64_t)stats.calls_executed());
+    yaml.key("calls_total").value((uint64_t)stats.calls_total());
     yaml.key("calls_fraction").value(stats.calls_fraction());
     yaml.key("calls_sort_fraction").value(stats.calls_sort_fraction());
     cov_stats_t::format_row_labels(stats.calls_by_status(), absbuf, pcbuf);
     yaml.key("calls_abslabel").value(absbuf);
     yaml.key("calls_pclabel").value(pcbuf);
 
-    yaml.key("branches_executed").value(stats.branches_executed());
-    yaml.key("branches_taken").value(stats.branches_taken());
-    yaml.key("branches_total").value(stats.branches_total());
+    yaml.key("branches_executed").value((uint64_t)stats.branches_executed());
+    yaml.key("branches_taken").value((uint64_t)stats.branches_taken());
+    yaml.key("branches_total").value((uint64_t)stats.branches_total());
     yaml.key("branches_fraction").value(stats.branches_fraction());
     yaml.key("branches_sort_fraction").value(stats.branches_sort_fraction());
     cov_stats_t::format_row_labels(stats.branches_by_status(), absbuf, pcbuf);
@@ -397,7 +397,7 @@ generate_annotated_source(const gghtml_params_t &params, cov_file_t *f,
 	yaml.key("status_long").value(cov::long_name(annotator.status()));
 	yaml.key("count").value(annotator.count());
 	yaml.key("count_if_instrumented").value(annotator.count_as_string());
-	yaml.key("lineno").value(annotator.lineno());
+	yaml.key("lineno").value((unsigned int)annotator.lineno());
 	yaml.key("blocks").value(annotator.blocks_as_string());
 	yaml.key("text").value(annotator.text());
 
@@ -461,7 +461,7 @@ generate_functions(const gghtml_params_t &params)
 	    yaml.key("url").value(url);
 	    yaml.key("status").value(cov::short_name(st));
 	    yaml.key("filename").value(loc->filename);
-	    yaml.key("lineno").value(loc->lineno);
+	    yaml.key("lineno").value((unsigned int)loc->lineno);
 	    yaml.key("stats"); generate_stats(yaml, stats);
 	    yaml.end_mapping();
 	}
